requirements.txt
requests
bs4
python-dotenv
ratelimit
gpt_researcher
agno==1.7.9
googlesearch-python
pycountry
fastapi